backups/psql-backup/requirements.txt
boto3
influxdb
PyYAML